% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary_S3methods.R
\name{as.character.R6}
\alias{as.character.R6}
\title{Create a string representation of an R6 class}
\usage{
\method{as.character}{R6}(x)
}
\arguments{
\item{x}{(\link[R6:R6Class]{R6::R6Class})}
}
\description{
Creates a string representation of an \link[R6:R6Class]{R6::R6Class} if not
already defined.
}
