% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary_S3methods.R
\name{[.Dictionary<-}
\alias{[.Dictionary<-}
\title{Set an element in a Dictionary}
\usage{
\method{[}{`Dictionary<-`}(object, i, value)
}
\arguments{
\item{object}{(\link{Dictionary})}

\item{i}{(\code{character()}) \cr
Keys of items to add.}

\item{value}{(\code{character()}) \cr
Values of items to add.}
}
\description{
This is simply a wrapper around \verb{[Dictionary]$add(i, value)}
}
\details{
Cannot be used to re-value an item.
}
